updated navbar styling
