+++ src/common/Navbar.js
@@ -1,5 +1,7 @@
 import React from 'react';
-import { StyleSheet, Text, View, TouchableOpacity } from 'react-native';
+import { Text, View, TouchableOpacity } from 'react-native';
+
+import * as styles from '../styles/navbarStyles';
 
 const Navbar = ({ header, goTo, back, plus }) => {
   return (
@@ -20,33 +22,5 @@ const Navbar = ({ header, goTo, back, plus }) => {
     </View>
   );
 };
-
-const styles = StyleSheet.create({
-  container: {
-    flexDirection: 'row',
-    width: '100%',
-    backgroundColor: 'white',
-    justifyContent: 'space-between',
-    paddingTop: 35,
-    paddingBottom: 30
-  },
-  text: {
-    color: '#444247',
-    fontSize: 32,
-    fontFamily: 'Avenir-Heavy'
-  },
-  back: {
-    color: '#444247',
-    fontSize: 32,
-    fontFamily: 'Avenir-Heavy',
-    marginLeft: 15
-  },
-  plus: {
-    color: '#444247',
-    fontSize: 32,
-    fontFamily: 'Avenir-Heavy',
-    marginRight: 15
-  }
-});
 
 export default Navbar;
